Make bulk importer optional.
Make BulkImporter work in two steps (initialize and import) instead of a
big initialize block.
Mark the file input as optiona, and make sure the controller make the
right thing with it, it does not change requirements if there is no
file. Also added 'csv' file requirement.
BulkImporter does not create the career anymore, it just do the import,
so the controller takes care of that.